requirements.txt
beautifulsoup4
ddgs
docling
duckduckgo-search>=6.0.0
email-validator
faiss-cpu
fastapi
google-generativeai
groq
langchain==0.2.16
langchain-community==0.2.16
langchain-core==0.2.38
langchain-google-genai
langchain-groq
langchain-huggingface==0.0.3
langchain-ollama
langchain-openai
lxml
markdown
openai
openpyxl
pandas
pydantic
pypdf
pytest
python-docx
python-dotenv
python-multipart
reportlab
requests
sentence-transformers
sqlalchemy
streamlit
tiktoken
uvicorn
watchdog
